refactor: 폴더명 recoil에서 globalState로 변경
